requirement.txt
scikit-learn
xgboost
mlflow
pandas
numpy
pyyaml
joblib
fastapi
uvicorn
streamlit